requirements.txt
# ReviewSignal 5.0 - Dependencies
# Production-ready requirements

# ═══════════════════════════════════════════════════════════════
# Core Framework
# ═══════════════════════════════════════════════════════════════
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic[email]==2.5.3
python-multipart==0.0.6

# ═══════════════════════════════════════════════════════════════
# Database
# ═══════════════════════════════════════════════════════════════
psycopg2-binary==2.9.9
asyncpg==0.29.0
sqlalchemy==2.0.25
alembic==1.13.1

# ═══════════════════════════════════════════════════════════════
# Cache & Queue
# ═══════════════════════════════════════════════════════════════
redis==5.0.1
aioredis==2.0.1

# ═══════════════════════════════════════════════════════════════
# Google Maps API
# ═══════════════════════════════════════════════════════════════
googlemaps==4.10.0

# ═══════════════════════════════════════════════════════════════
# Web Scraping (for Google Maps, Yelp, TripAdvisor, Trustpilot)
# ═══════════════════════════════════════════════════════════════
playwright==1.41.0
beautifulsoup4==4.12.3
lxml==5.1.0
httpx==0.26.0
aiohttp==3.9.3

# ═══════════════════════════════════════════════════════════════
# Lead Generation - Apollo.io API (ZAMIAST LinkedIn)
# ═══════════════════════════════════════════════════════════════
# Apollo.io jest używany przez zewnętrzne API - brak dedykowanej biblioteki Python
# Integracja przez REST API (httpx już jest zainstalowany)

# ═══════════════════════════════════════════════════════════════
# Machine Learning
# ═══════════════════════════════════════════════════════════════
numpy==1.26.3
scipy==1.12.0
scikit-learn==1.4.0
pandas==2.2.0
vaderSentiment==3.3.2  # VADER sentiment scoring (local, zero API cost)

# ═══════════════════════════════════════════════════════════════
# Neural Core - Semantic Embeddings (MiniLM, 80MB, CPU-friendly)
# ═══════════════════════════════════════════════════════════════
sentence-transformers==2.2.2
# torch is installed automatically with sentence-transformers (CPU version)

# ═══════════════════════════════════════════════════════════════
# Payments
# ═══════════════════════════════════════════════════════════════
stripe==7.12.0

# ═══════════════════════════════════════════════════════════════
# Authentication
# ═══════════════════════════════════════════════════════════════
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# ═══════════════════════════════════════════════════════════════
# Logging & Monitoring
# ═══════════════════════════════════════════════════════════════
structlog==24.1.0
rich==13.7.0
prometheus-client==0.19.0

# ═══════════════════════════════════════════════════════════════
# AI / LLM - Autonomous Agent
# ═══════════════════════════════════════════════════════════════
anthropic==0.76.0  # Claude Opus 4.5, Sonnet 4.5, Haiku 4.5

# ═══════════════════════════════════════════════════════════════
# PDF Generation
# ═══════════════════════════════════════════════════════════════
reportlab==4.0.9
Pillow==10.2.0  # For image handling in PDFs

# ═══════════════════════════════════════════════════════════════
# Email Delivery
# ═══════════════════════════════════════════════════════════════
resend>=1.0.0

# ═══════════════════════════════════════════════════════════════
# Utilities
# ═══════════════════════════════════════════════════════════════
python-dotenv==1.0.0
tenacity==8.2.3
tqdm==4.66.1
